Redesign CLI structure and configuration management (#146)
* Initial commit

* Go mod

* Go mod

* Update types

* Update merge

* Update merge

* Update export

* Create setupCluster function

* Clean

* Bee export with reflection

* NodeGroup export with reflection

* Use new config in checks

* Temp

* Temp

* Clean up helpers

* Bump version

* Refactor ping

* Check refactor fondation

* Go mod tidy

* New checks fondation for refactor

* Refactor ping

* Refactor ping

* Remove ping cmd

* Refactor Kademlia

* Refactor Kademlia

* Refactor Kademlia

* Add comment

* Fix bee pod

* Support image pull secret

* Update .gitignore

* Refactor balances

* Refactor balances

* Refactor chunkrepair

* Refactor chunkrepair

* Refactor file-retrieval

* Refactor full-connectivity

* Refactor full-connectivity

* Refactor gc

* Refactor localpinning

* Refactor localpinning

* Refactor localpinning

* Refactor manifest

* Refactor peercount

* Refactor pingpong

* Refactor pingpong

* Refactor pss

* Refactor pingpong

* Refactor pullsync

* Improve localpinning ctx

* Refactor pushsync

* Refactor retrieval

* Refactor settlements

* Refactor soc

* Refactor check cmd

* Refactor

* Implement default options for checks

* Improve Run[profile] selection

* Initial apply options

* Initial apply options

* Use constructor for balances default options

* Implement apply options function

* Rename RunProfiles to Playbooks

* Move checks to config

* Redesign config file

* Redesign config file

* Optimize applying options

* Extract global check options

* Set cluster globaly

* Update checks

* Rewrite stress to simulation

* Clean start and delete commands

* Rewrite print cmd

* Remove unused cmds and pkgs

* Remove temp code

* Rewrite create and delete cmds

* Implement runner pkg

* Rename runner pkg to beekepeer

* Rename

* Rename

* Rename

* Fix typo

* Refactor config inheritence

* Update comments

* Refactor cmd

* Move k8sClient into command

* Flatten cluster config

* Add inherit

* Wire flags

* Implement cluster inheritance

* Fix flags

* Read dir for all config files

* Add support for light nodes

* Small refactor

* Separate helm cluster

* Rename function

* Fix pss metrics

* Align check options

* go mod tidy

* Bump version

* Mark checks for update

* Refactor k8s cmds

* Improve config dir reading

* Fix lint

* Add comments to cmd

* Add comments to config

* Add description to config file

* Add option to delete storage

* Use random seed by default

* Refactor

* Simplify err return

* Fix description

* Update comment

* Fix misspell

* Update config types

* Fix command nesting

* Support kubernetes 1.19

* Support kubernetes 1.19

* Update config types

* Initial commit

* Swap service outline

* Swap service outline

* Swap service outline

* Implement swap

* Fund update

* go mod tidy

* big int

* Update sw3 bindings

* Refactor

* Update funding

* Update funding

* Create swap interface

* Update swap constants

* Update bee config

* Fix bzz sending

* Update deposits

* Update light node image settings

* Add retries to fund

* Update node count

* go mod tidy

* Update funding outputs

* Update ingress to new k8s api

* Implement swap noop

* Implement swap noop

* Create preRunE method

* Update cmd timeouts

* Add comment

* Add gas price to funding

* Use float6 for funding

* Misspell

* Misspell

* Update config to support central geth

* Update README and descriptions

* Fix swap

* Use strings.TrimLeft

* Use strings.TrimPrefix

* Remove prints

* Use bigCoin.SetInt64

* Add config for ci and local cluster

* Fix balances check

* Fix lightnode group name

* Make funding optional

* add tested ci.yaml config

* add default .beekeeper.yaml for local tests

* set default kubeconfig

* rename ci.yaml to local.yaml and set dns and gc configs

* fix undefined node groups

* only one lvl of inheritance

* increase file for settlements test to 2mb

* increase postage stamp depth for settlements local test

* Make clef settings uniform for all nodes

* Disable clef by default

* add local-clef cluster

* rename wait-before-downloading to wait-before-download

* add settlements test for clef

* Change WaitBeforeDownload type

* Change WaitBeforeDownload type

* increase wait for settlements test for clef

* fix: adjust address-prefix

* apply pss config fix on local.yaml

+++ pkg/k8s/namespace/namespace.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/ethersphere/beekeeper"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -29,12 +30,16 @@ type Options struct {
 }
 
 // Create creates namespace
-func (c *Client) Create(ctx context.Context, name string, o Options) (err error) {
+func (c *Client) Create(ctx context.Context, name string) (err error) {
 	spec := &v1.Namespace{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:        name,
-			Annotations: o.Annotations,
-			Labels:      o.Labels,
+			Name: name,
+			Annotations: map[string]string{
+				"created-by": fmt.Sprintf("beekeeper:%s", beekeeper.Version),
+			},
+			Labels: map[string]string{
+				"app.kubernetes.io/managed-by": "beekeeper",
+			},
 		},
 	}
 
@@ -58,6 +63,16 @@ func (c *Client) Update(ctx context.Context, name string, o Options) (err error)
 
 // Delete deletes namespace
 func (c *Client) Delete(ctx context.Context, name string) (err error) {
+	n, err := c.clientset.CoreV1().Namespaces().Get(ctx, name, metav1.GetOptions{})
+	if err != nil {
+		return err
+	}
+
+	managedBy, ok := n.Labels["app.kubernetes.io/managed-by"]
+	if !ok || managedBy != "beekeeper" {
+		return fmt.Errorf("namespace %s is not managed by beekeeper, try kubectl", name)
+	}
+
 	err = c.clientset.CoreV1().Namespaces().Delete(ctx, name, metav1.DeleteOptions{})
 	if err != nil {
 		if errors.IsNotFound(err) {
